Refactor AdvancedSelectQueryBuilder component
